app/requirements-lite.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
prometheus_client==0.20.0
Pillow==10.4.0
numpy==1.26.4
